Fix MSVC warnings

+++ 3rdparty/SDRPlusPlus/dsp/types.h
@@ -51,7 +51,7 @@ namespace dsp {
         }
 
         inline float amplitude() {
-            return sqrt((re * re) + (im * im));
+            return sqrtf((re * re) + (im * im));
         }
 
         inline float fastAmplitude() {
